feat(suite-native): Implement home stack navigation
