+++ packages/suite-native-home/src/navigation/routes.ts
@@ -1,0 +1,9 @@
+export enum HomeStackRoutes {
+    Home = 'Home',
+    HomeDetail = 'HomeDetail',
+}
+
+export type HomeStackParamList = {
+    [HomeStackRoutes.Home]: undefined;
+    [HomeStackRoutes.HomeDetail]: { message: string };
+};
