+++ packages/suite-native-navigation-root/src/navigation/routes.ts
@@ -1,7 +1,10 @@
+import { NavigatorScreenParams } from '@react-navigation/native';
+import { HomeStackParamList } from '@suite-native/home';
+
 import { TabsOption } from '@suite-native/navigation';
 
 export enum RouteTabs {
-    Home = 'Home',
+    HomeStack = 'HomeStack',
     Accounts = 'Accounts',
     Action = 'Action',
     Prices = 'Prices',
@@ -9,7 +12,7 @@ export enum RouteTabs {
 }
 
 export type RootTabsParamList = {
-    [RouteTabs.Home]: undefined;
+    [RouteTabs.HomeStack]: NavigatorScreenParams<HomeStackParamList>;
     [RouteTabs.Accounts]: undefined;
     [RouteTabs.Action]: undefined;
     [RouteTabs.Prices]: undefined;
@@ -17,7 +20,7 @@ export type RootTabsParamList = {
 };
 
 export const rootTabsOptions: TabsOption = {
-    [RouteTabs.Home]: {
+    [RouteTabs.HomeStack]: {
         iconName: 'home',
     },
     [RouteTabs.Accounts]: {
